fix: allow UI notifications when filter text returns non-zero list
this fixes the bug where when the user types a filter that shows no data, (e.g. "f" in smallcities.json) and then clears the query, the UI still shows "no data" instead of now showing the full list. This is because we are only notifying UI when there's no data, but not notifying the UI when there's data